retro class, fragment can pull info
