--- app/src/main/java/com/example/jello/afterschool/fragments/StudentsFragment.java
@@ -1,23 +1,0 @@
-package com.example.jello.afterschool.fragments;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.example.jello.afterschool.R;
-
-/**
- * Created by jello on 1/27/17.
- */
-
-public class StudentsFragment extends Fragment {
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.student_view, container, false);
-    }
-}
