+++ app/src/main/java/com/example/jello/afterschool/model/Children.java
@@ -1,0 +1,57 @@
+package com.example.jello.afterschool.model;
+
+import java.util.List;
+
+public class Children {
+
+    private List<String> classrooms;
+
+    private String childId;
+
+    private String name;
+
+    private String parent;
+
+    private String image;
+
+    public List<String> getClassrooms() {
+        return classrooms;
+    }
+
+    public void setClassrooms(List<String> classrooms) {
+        this.classrooms = classrooms;
+    }
+
+    public String getChildId() {
+        return childId;
+    }
+
+    public void setChildId(String childId) {
+        this.childId = childId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getParent() {
+        return parent;
+    }
+
+    public void setParent(String parent) {
+        this.parent = parent;
+    }
+
+    public String getImage() {
+        return image;
+    }
+
+    public void setImage(String image) {
+        this.image = image;
+    }
+
+}
